+++ plugins/experimental/gzip/misc.cc
@@ -61,9 +61,9 @@ normalize_accept_encoding(TSHttpTxn txnp, TSMBuffer reqp, TSMLoc hdr_loc)
         --value_count;
         val = TSMimeHdrFieldValueStringGet(reqp, hdr_loc, field, value_count, &val_len);
 
-        if (val_len == strlen("gzip"))
+        if (val_len == (int) strlen("gzip"))
           gzip = !strncmp(val, "gzip", val_len);
-        else if (val_len == strlen("deflate"))
+        else if (val_len == (int) strlen("deflate"))
           deflate = !strncmp(val, "deflate", val_len);
       }
     }
